Short: New libhttpd and all atytending changes Long:
Libhttpd's old httpd struct is now split in httpd for server config and
request for request specific stuff. This needs testing and some tidying
up.

Wifidog has been altered accordingly. There's now a child thread created to
serve ALL requests. No more iterating-until-we-spawn-a-thread-and-then-
pretending-we're-done-when-we're-really-not hack. A few more tidbits.

This all needs *SERIOUS* testing.