Fix missing error recovery in libhttpd/api.c
The select call sometimes may be interrupted or otherwise return
a value < 0. In this case, we do want to signal an error to the caller.

For subsequent calls, the error need to be reset or the caller (gateway.c)
will keep going into the error handler path even though the HTTP request
was handled succesfully.