fix the bug of getting the valule of Gateway mistakenly from the valule of Destination
The /proc/net/route info on the device is always displayed like this:

$cat /proc/net/route
Iface   Destination   Gateway     Flags   RefCnt  Use  Metric  Mask      MTU  Window  IRTT
eth0    00A6440A      00000000     0001   0       0    0       00FFFFFF  0    0       0

clearly, the 3rd vaule(Gateway) is what we want, but in the code as we can see,
we always get the 2nd value(Destination), that's not right.

I think the reason we didn't recognize this bug is that sometimes Destination can be '000000',
but recently, a "failed to detect the external interface" err occured on my device , just cause the
value of Destination is not "00000000"(but Gateway is "00000000", means this err should not happen).
hope can help to fix this bug, thanks:)

Signed-off-by: hongze zhu <hongze.zhu@gmail.com>